+++ flinkboot-core/src/main/java/io/github/sekelenao/flinkboot/core/internal/resource/ClasspathResource.java
@@ -1,11 +1,12 @@
 package io.github.sekelenao.flinkboot.core.internal.resource;
 
 import io.github.sekelenao.flinkboot.core.api.exception.resource.ResourceNotFoundException;
+import io.github.sekelenao.flinkboot.core.api.resource.Resource;
 
 import java.io.InputStream;
 import java.util.Objects;
 
-class ClasspathResource implements Resource {
+public class ClasspathResource implements Resource {
 
     private final String location;
 
@@ -15,14 +16,22 @@ class ClasspathResource implements Resource {
 
     @Override
     public InputStream inputStream() {
-        Objects.requireNonNull(location);
-        var cleanPath = location.startsWith("/") ? location.substring(1) : location;
+        var cleanPath = location;
+        while (cleanPath.startsWith("/")) {
+            cleanPath = cleanPath.substring(1);
+        }
         var classLoader = Thread.currentThread().getContextClassLoader();
         if (classLoader == null) {
             classLoader = ClasspathResource.class.getClassLoader();
         }
+        if (classLoader == null) {
+            classLoader = ClassLoader.getSystemClassLoader();
+        }
+        if (classLoader == null) {
+            throw new ResourceNotFoundException(location);
+        }
         var stream = classLoader.getResourceAsStream(cleanPath);
-        if(stream == null){
+        if (stream == null) {
             throw new ResourceNotFoundException(location);
         }
         return stream;